requirements.txt
transformers>=4.30.0
torch
python-chess
accelerate
